requirements.txt
# LangGraph y LangChain
langgraph>=1.0.7
langchain>=1.2.6
langchain-core>=1.2.7
langchain-openai>=1.1.7
langchain-community>=0.4.0

# Utilidades
python-dotenv>=1.0.0
typing-extensions>=4.8.0
httpx>=0.27.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright==1.57.0
ddgs>=9.0.0

# Opcional: para visualización
matplotlib>=3.10.8

# Tests
pytest>=8.0.0
pytest-asyncio>=0.23.0
